Update code for pay_period table
The data type had to be changed in the ETL.py file so that the postfres could read in the year. It doesn't allow for year to be an np.int. I also updated the sql_queries.py file becuase it had an error (wrote "per" instead of "pay" on the insert statement.)