fix: insert to payroll_type table code
Previously when we inserted the payroll_type data, the code would do
nothing if the data was repeated. Which led to weird unique key ids. For
example, if rows 1 was labeled 'SA' ,row 2 was labeled 'SA', and row 3
was labeled 'RA'. In the data base 'SA' would have a unique ID of 1 but
'RA' would have a unique id of 3, which corresponded to the row number
instead of the order it was entered. Now the code inserts the unique key
id with an entry order.